% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ae_factors_to_numeric}
\alias{ae_factors_to_numeric}
\title{Illustration of crayon colors}
\usage{
ae_factors_to_numeric(x)
}
\arguments{
\item{x}{A string of numeric factors, which need to be transformed to numeric}
}
\value{
Numeric values of the factor
}
\description{
Creates a plot of the crayon colors in \code{\link{brocolors}}
}
\examples{
to_numeric(factor(c(1,2,3,43,7)))

}
