% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ae_get_gs_data}
\alias{ae_get_gs_data}
\title{GoogleSheets to dataframe}
\usage{
ae_get_gs_data(worksheet, sheet_key, colnames = TRUE, path_to_token = NULL)
}
\arguments{
\item{worksheet}{The worksheet in the googlesheets, which you want to retrieve data from}

\item{sheet_key}{The sheet_key of the worksheet. Can be accessed through the URL}

\item{colnames}{Boolean on whether you wish to have the first row of the spreadsheet serve as column names}

\item{path_to_token}{Specify path to token, which allows access to googlesheets. If not specified, you
will be asked to grant access via a web browser}
}
\value{
Tibble of googlesheets data
}
\description{
Retrieves data from a googlesheets online, if you have authorization
}
\examples{
ae_get_gs_data("your_data","your_worksheet_name","your_sheet_key_id_like_1zmxDhJ...")

}
