% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ae_tex_add_block_img}
\alias{ae_tex_add_block_img}
\title{Latex Text Block}
\usage{
ae_tex_add_block_img(
  path_to_image,
  x_cor,
  y_cor,
  textbox_width = "13.25",
  angle = 0
)
}
\arguments{
\item{x_cor}{Specify the \code{x_cor} in cm for the top left corner of the textbox}

\item{y_cor}{Specify the \code{y_cor} in cm for the top left corner of the textbox}

\item{text}{Specify the text you want to add to the Latex document}

\item{font_size}{Specify the \code{font_size} in points. I.e. "12".}

\item{font}{Specify the \code{font}. Accepted arguments include "light", "medium" and "bold.}

\item{colnames}{Boolean on whether you wish to have the first row of the spreadsheet serve as column names}

\item{path_to_token}{Specify path to token, which allows access to googlesheets. If not specified, you
will be asked to grant access via a web browser}
}
\value{
Tibble of googlesheets data
}
\description{
Add textblock to latex document via R function
}
\examples{
ae_get_gs_data("your_data","your_worksheet_name","your_sheet_key_id_like_1zmxDhJ...")

}
